requirements.txt
scipy~=1.9.3
numpy~=1.21.5
opencv-python>=4.0.0
flask~=1.1.2
pyserial>=3.5
json
datetime
requests~=2.26.0